feat(data): modify the playerData to ModData
